backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
pgvector>=0.2.4

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# LLM
llama-cpp-python==0.2.90

langgraph>=0.0.26
duckduckgo-search>=5.0.0

# Vector Store
sentence-transformers>=2.3.0

# Document Processing
python-docx>=1.1.0
python-pptx>=0.6.23
PyPDF2>=3.0.0
openpyxl>=3.1.2
chardet>=5.2.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
psutil>=5.9.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
